[new]: Program Development by Stepwise Refinement
By Niklaus Wirth, 1971

This article very nicely demonstrates a core idea in Computer Science,
that through experimentation and observation along with applied logic
can we find more and more optimal solutions for any given problem.

It also touches on a problem software still faces to day which is the
issue of adapability, something we now call maintainablity or portability.
That by coding ourselves into a corner we make software impossible to
improve without completely rewriting it.  Being mindful of change allows
us to build systems that can easily rise to any challenge.